+++ packages/granularity/src/components/GrDropdownMenu/GrDropdownMenuColumns.vue
@@ -1,28 +1,20 @@
 <script setup lang="ts">
-import { computed } from 'vue'
+import { colsClass, columnsBaseClass, type GrDropdownMenuColumnsCount } from './grDropdownMenuStyles'
 
 export interface GrDropdownMenuColumnsProps {
-  cols?: 2 | 3 | 4
+  cols?: GrDropdownMenuColumnsCount
 }
 
-const props = withDefaults(defineProps<GrDropdownMenuColumnsProps>(), {
+withDefaults(defineProps<GrDropdownMenuColumnsProps>(), {
   cols: 2,
-})
-
-const colsClass = computed(() => {
-  return {
-    2: 'grid-cols-2',
-    3: 'grid-cols-3',
-    4: 'grid-cols-4',
-  }[props.cols]
 })
 </script>
 
 <template>
   <div
     data-gr-dropdown-menu-columns
-    class="w-full grid divide-x divide-[var(--gr-brd)]"
-    :class="colsClass"
+    role="none"
+    :class="[columnsBaseClass, colsClass[cols]]"
   >
     <slot />
   </div>