+++ packages/granularity/src/components/GrDropdownMenu/GrDropdownMenuDivider.vue
@@ -1,4 +1,6 @@
 <script setup lang="ts">
+import { dividerClass, dividerInsetClass } from './grDropdownMenuStyles'
+
 export interface GrDropdownMenuDividerProps {
   inset?: boolean
 }
@@ -12,7 +14,6 @@ withDefaults(defineProps<GrDropdownMenuDividerProps>(), {
   <div
     data-gr-dropdown-menu-divider
     role="separator"
-    class="border-t border-[var(--gr-brd)]"
-    :class="inset ? 'mx-2' : ''"
+    :class="[dividerClass, inset ? dividerInsetClass : '']"
   />
 </template>