+++ packages/granularity/src/components/GrDropdownMenu/config.ts
@@ -1,6 +1,9 @@
 import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
 
+import { grDropdownMenuSafelist } from './safelist'
+
 export const grDropdownMenuConfig = defineGranularComponent(import.meta.url, {
   name: 'GrDropdownMenu',
   dependencies: ['GrDropdown'],
+  safelist: grDropdownMenuSafelist,
 })